#6 [refactor] UseCase 위치 수정
