+++ domain/src/main/java/com/sopetit/domain/base/UseCase.kt
@@ -1,4 +1,4 @@
-package com.sopetit.domain
+package com.sopetit.domain.base
 
 import kotlinx.coroutines.flow.Flow
